requirements.txt
grpcio-tools
pydantic